requirements.txt
pandas
openai
psycopg
python-dotenv
timescale-vector
langchain
google-genai
opencv-python

numpy
requests
langchain-community==0.3.0
langchain-chroma==0.1.4
langchain-ollama==0.2.0

pdf2image==1.17.0

flask
flask-cors

ipykernel
matplotlib
